pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "VERSUS project"
authors = ["TEAM DR.Bigyo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
django-redis = "^5.4.0"
gunicorn = "^21.2.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.3"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
